create Header component, parent of PromotionBanner and SearchBar. add components in separate folders
