+++ client/src/App.js
@@ -1,18 +1,17 @@
-import TopBar from './components/TopBar'
-import PromotionBanner from './components/PromotionBanner'
-import SearchBar from './components/SearchBar'
-import ProductCategories from './components/ProductCategories'
-import BottomBar from './components/BottomBar'
+import TopBar from './components/NavBar/TopBar'
+import Header from './components/Header/Header';
+import ProductCategories from './components/Products/ProductCategories'
+import BottomBar from './components/Footer/BottomBar'
 
 function App() {
   return (
     <div className="flex flex-col bg-neutral-800 h-screen w-screen">
       <header>
         <TopBar />
-        <PromotionBanner />
+
       </header>
       <main>
-        <SearchBar />
+        <Header />
         <ProductCategories />
       </main>
       <footer>
