+++ client/src/components/Footer/BottomBar.js
